+++ android/src/main/kotlin/top/kikt/imagescanner/ImageScanner.kt
@@ -32,6 +32,8 @@ class ImageScanner(val registrar: PluginRegistry.Registrar) {
             MediaStore.Images.Media.TITLE, // id
             MediaStore.Images.Media.BUCKET_ID, // dir id 目录
             MediaStore.Images.Media.BUCKET_DISPLAY_NAME, // dir name 目录名字
+            MediaStore.Images.Media.WIDTH, // 宽
+            MediaStore.Images.Media.HEIGHT, // 高
             MediaStore.Images.Media.DATE_TAKEN //日期
     )
 
@@ -45,11 +47,19 @@ class ImageScanner(val registrar: PluginRegistry.Registrar) {
             MediaStore.Video.Media.BUCKET_ID, // dir id 目录
             MediaStore.Video.Media.BUCKET_DISPLAY_NAME, // dir name 目录名字
             MediaStore.Video.Media.DATE_TAKEN, //日期
+            MediaStore.Video.Media.WIDTH, // 宽
+            MediaStore.Video.Media.HEIGHT, // 高
             MediaStore.Video.Media.DURATION //时长
     )
 
 
     private var imgList = ArrayList<Img>()
+    /// dirId,ImgList
+    private val map = HashMap<String, ArrayList<Img>>()
+    private val idPathMap = HashMap<String, String>()
+    private val pathIdMap = HashMap<String, String>()
+    private var thumbMap = HashMap<String, String>()
+    private val pathImgMap = HashMap<String, Img>()
 
     private fun scan() {
         Log.i("K", "start scan")
@@ -91,7 +101,9 @@ class ImageScanner(val registrar: PluginRegistry.Registrar) {
             val thumb = mCursor.getString(mCursor.getColumnIndex(MediaStore.Images.ImageColumns.MINI_THUMB_MAGIC))
             val imgId = mCursor.getString(mCursor.getColumnIndex(MediaStore.Images.ImageColumns._ID))
             val date = mCursor.getLong(mCursor.getColumnIndex(MediaStore.Images.ImageColumns.DATE_TAKEN))
-            val img = Img(path, imgId, dir, dirId, title, thumb, AssetType.Image, date, null)
+            val width = mCursor.getInt(mCursor.getColumnIndex(MediaStore.Images.Media.WIDTH))
+            val height = mCursor.getInt(mCursor.getColumnIndex(MediaStore.Images.Media.HEIGHT))
+            val img = Img(path, imgId, dir, dirId, title, thumb, AssetType.Image, date, null, width, height)
 
             val file = File(path)
             if (file.exists().not()) {
@@ -132,7 +144,11 @@ class ImageScanner(val registrar: PluginRegistry.Registrar) {
             val imgId = mCursor.getString(mCursor.getColumnIndex(MediaStore.Video.Media._ID))
             val date = mCursor.getLong(mCursor.getColumnIndex(MediaStore.Video.Media.DATE_TAKEN))
             val durationMs = mCursor.getLong(mCursor.getColumnIndex(MediaStore.Video.Media.DURATION))
-            val img = Img(path, imgId, dir, dirId, title, thumb, AssetType.Video, date, durationMs)
+
+            val width = mCursor.getInt(mCursor.getColumnIndex(MediaStore.Video.Media.WIDTH))
+            val height = mCursor.getInt(mCursor.getColumnIndex(MediaStore.Video.Media.HEIGHT))
+
+            val img = Img(path, imgId, dir, dirId, title, thumb, AssetType.Video, date, durationMs, width, height)
 
             val file = File(path)
             if (file.exists().not()) {
@@ -157,12 +173,6 @@ class ImageScanner(val registrar: PluginRegistry.Registrar) {
             result?.success(map.keys.toList())
         }
     }
-
-    /// dirId,ImgList
-    val map = HashMap<String, ArrayList<Img>>()
-
-    val idPathMap = HashMap<String, String>()
-    val pathIdMap = HashMap<String, String>()
 
     private fun split() {
         map.clear()
@@ -311,8 +321,6 @@ class ImageScanner(val registrar: PluginRegistry.Registrar) {
         }
     }
 
-    var thumbMap = HashMap<String, String>()
-
     private fun scanThumb() {
         val mImageUri = MediaStore.Images.Thumbnails.EXTERNAL_CONTENT_URI
         val mContentResolver = registrar.activity().contentResolver
@@ -332,8 +340,6 @@ class ImageScanner(val registrar: PluginRegistry.Registrar) {
         val path = call.arguments as String
     }
 
-    val pathImgMap = HashMap<String, Img>()
-
     private fun getThumbFromPath(img: Img?): String? {
         if (img == null) {
             return null
@@ -344,7 +350,7 @@ class ImageScanner(val registrar: PluginRegistry.Registrar) {
     fun getImageThumb(call: MethodCall, result: MethodChannel.Result) {
         threadPool.execute {
             val path = call.arguments as String
-            val img = pathImgMap[path]
+            val img = getImageWithId(path)
             if (img == null) {
                 result.success(null)
             } else {
@@ -364,7 +370,7 @@ class ImageScanner(val registrar: PluginRegistry.Registrar) {
 
         val args = call.arguments as List<Any>
         val id = args[0] as String
-        val img = pathImgMap[id] ?: return
+        val img = getImageWithId(id) ?: return
         val width = (args[1] as String).toInt()
         val height = (args[2] as String).toInt()
 
@@ -382,7 +388,7 @@ class ImageScanner(val registrar: PluginRegistry.Registrar) {
         val resultList = ArrayList<String>()
 
         idList.forEach { id ->
-            val img = pathImgMap[id]
+            val img = getImageWithId(id)
             img?.apply {
                 resultList.add(typeFromEntity(this))
             }
@@ -400,7 +406,7 @@ class ImageScanner(val registrar: PluginRegistry.Registrar) {
 
     fun getAssetDurationWithId(call: MethodCall, result: MethodChannel.Result) {
         val id = call.arguments<String>()
-        val img = pathImgMap[id]
+        val img = getImageWithId(id)
         if (img == null || img.type != AssetType.Video) {
             result.success(null)
         } else {
@@ -411,6 +417,33 @@ class ImageScanner(val registrar: PluginRegistry.Registrar) {
                 result.success(duration / 1000)
             }
         }
+    }
+
+    private fun getImageWithId(id: String) = pathImgMap[id]
+
+    fun getSizeWithId(call: MethodCall, result: MethodChannel.Result) {
+        val id = call.arguments<String>()
+        val img = getImageWithId(id)
+        if (img == null) {
+            result.success(mapOf<String, Int>())
+            return
+        }
+        val sizeMap = mapOf(
+                "width" to img.width,
+                "height" to img.height
+        )
+        result.success(sizeMap)
+    }
+
+    fun releaseMemCache(result: MethodChannel.Result) {
+        imgList.clear()
+        idPathMap.clear()
+        pathIdMap.clear()
+        pathImgMap.clear()
+        map.clear()
+        thumbMap.clear()
+
+        result.success(1)
     }
 }