+++ example/lib/photo_list.dart
@@ -1,3 +1,4 @@
+import 'dart:async';
 import 'dart:io';
 import 'dart:typed_data';
 
@@ -28,21 +29,25 @@ class _PhotoListState extends State<PhotoList> {
 
   Widget _buildItem(BuildContext context, int index) {
     AssetEntity entity = widget.photos[index];
-    print("request index = $index , image id = ${entity.id} type = ${entity.type}");
+    print(
+        "request index = $index , image id = ${entity.id} type = ${entity.type}");
 
-    Future<Uint8List> thumbDataWithSize = entity.thumbDataWithSize(500, 500); // get thumb with width and height.
-    Future<Uint8List> thumbData = entity.thumbData; // the method will get thumbData is size 64*64.
+    Future<Uint8List> thumbDataWithSize =
+        entity.thumbDataWithSize(500, 500); // get thumb with width and height.
+    Future<Uint8List> thumbData =
+        entity.thumbData; // the method will get thumbData is size 64*64.
     Future<Uint8List> imageFullData = entity.fullData; // get the origin data.
     Future<File> file = entity.file; // get file
     Future<Duration> length = entity.videoDuration;
-    length.then((v){
+    length.then((v) {
       print("duration = $v");
     });
 
     return FutureBuilder<Uint8List>(
       future: thumbDataWithSize,
       builder: (BuildContext context, AsyncSnapshot<Uint8List> snapshot) {
-        if (snapshot.connectionState == ConnectionState.done && snapshot.data != null) {
+        if (snapshot.connectionState == ConnectionState.done &&
+            snapshot.data != null) {
           return InkWell(
             onTap: () => showInfo(entity),
             child: Stack(
@@ -77,7 +82,11 @@ class _PhotoListState extends State<PhotoList> {
     if (entity.type == AssetType.video) {
       var file = await entity.file;
       var length = file.lengthSync();
-      print("${entity.id} length = $length");
+      var size = await entity.size;
+      print("${entity.id} length = $length , size = $size");
+    } else {
+      var size = await entity.size;
+      print("${entity.id} size = $size");
     }
   }
 }