+++ ios/Classes/ImageScanner.h
@@ -36,6 +36,9 @@
 
 - (void)isCloudWithCall:(FlutterMethodCall *)call result:(FlutterResult)result;
 
--(void)getDurationWithId:(FlutterMethodCall *)call result:(FlutterResult)result;
+- (void)getDurationWithId:(FlutterMethodCall *)call result:(FlutterResult)result;
 
+- (void)getSizeWithId:(FlutterMethodCall *)call result:(FlutterResult)result;
+
+- (void)releaseMemCache:(FlutterMethodCall *)call result:(FlutterResult)result;
 @end